+++ pkg/analysis_server/lib/src/utilities/import_analyzer.dart
@@ -181,7 +181,7 @@ class _ReferenceFinder extends RecursiveAstVisitor<void> {
 
   /// Initialize a newly created finder to send information to the [recorder].
   _ReferenceFinder(this.unit, this.recorder) {
-    for (var import in unit.libraryElement2.firstFragment.libraryImports2) {
+    for (var import in unit.libraryElement2.firstFragment.libraryImports) {
       _importsByPrefix
           .putIfAbsent(import.prefix2?.element.name ?? '', () => {})
           .add(import);